Added the CoST baseline

+++ baselines/src/models/costencoder.py
@@ -109,7 +109,8 @@ class CoSTEncoder(nn.Module):
         )
 
     def forward(self, x, tcn_output=False, mask='all_true'):  # x: B x T x input_dims
-       
+        x = x.float()
+        
         nan_mask = ~x.isnan().any(axis=-1)
         x[~nan_mask] = 0
         x = self.input_fc(x)  # B x T x Ch
@@ -162,5 +163,5 @@ class CoSTEncoder(nn.Module):
             out = mod(x)  # b t d
             season.append(out)
         season = season[0]
-        
+
         return trend, self.repr_dropout(season)
